Compilation fix
Former-commit-id: 26060e343d554b5da6a2cb75001fd6a1c5ae963b

+++ libraries/dimred/laplacianEigenmaps.cpp
@@ -48,5 +48,5 @@ void LaplacianEigenmap::reduceDimensionality()
 		MAT_ELEM(D,i,i)+=MAT_ELEM(G,i,j);
 	//Construct eigenmaps
 	generalizedEigs(L,D,mappedX,Y);
-	keepColumns(Y,1,outputDim);
+	keepColumns(Y,1,(int)outputDim);
 }
